app/requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==2.3.2
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.4.0
PyYAML==6.0.2
Werkzeug==3.0.4
zipp==3.20.2
